requirements.txt
fastapi[all]
nltk
pandas
gunicorn
bs4
requests
python-dotenv
psycopg2